SyncScreen
Setup a sync screen so that users who login or create without datastore being setup load for a bit first
