+++ src/screens/RootStackNavigator.tsx
@@ -2,10 +2,10 @@ import { createNativeStackNavigator } from "@react-navigation/native-stack";
 
 import HomeScreen from "./HomeScreen";
 import DrawerNav from "./drawer/DrawerNav";
+import SyncScreen from "./SyncScreen";
 import SpinningIndicator from "../components/SpinningIndicator";
 import { RootStackParamList } from "../types/NavigatorTypes";
 import { useLoad } from "../helper/LoadingContext";
-import Temp from "./temp";
 
 // Create a stack navigator to handle navigation throughout the app
 const Stack = createNativeStackNavigator<RootStackParamList>();
@@ -30,7 +30,11 @@ export function RootStackNavigator() {
           component={DrawerNav}
           options={{ headerShown: false }}
         />
-        <Stack.Screen name="Temp" component={Temp} />
+        <Stack.Screen
+          name="SyncScreen"
+          component={SyncScreen}
+          options={{ headerShown: false }}
+        />
       </Stack.Navigator>
       {isLoading ? <SpinningIndicator /> : null}
     </>
