+++ src/types/ContextTypes.ts
@@ -29,6 +29,4 @@ export type UserContextType = {
 export type LoadingContextType = {
   isLoading: boolean;
   setIsLoading: Dispatch<SetStateAction<boolean>>;
-  dataStoreReady: boolean;
-  setDataStoreReady: Dispatch<SetStateAction<boolean>>;
 };
